pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.3",
    "joblib>=1.5.3",
    "numpy>=2.4.4",
    "psycopg2-binary>=2.9.11",
    "scikit-learn>=1.8.0",
]
